+++ src/types/listen-utils.ts
@@ -2,12 +2,12 @@ import { Lambda, once } from "../utils/utils"
 import { untrackedStart, untrackedEnd } from "../core/derivation"
 
 export interface IListenable {
-    changeListeners: Function[] | null
+    changeListeners: Function[] | undefined
     observe(handler: (change: any, oldValue?: any) => void, fireImmediately?: boolean): Lambda
 }
 
 export function hasListeners(listenable: IListenable) {
-    return listenable.changeListeners && listenable.changeListeners.length > 0
+    return listenable.changeListeners !== undefined && listenable.changeListeners.length > 0
 }
 
 export function registerListener<T>(listenable: IListenable, handler: Function): Lambda {